removed scores from user router and placed them in score router
